Extract PageContent component from Page to avoid any confusion.
