+++ packages/components/src/PageContent/README.md
@@ -1,0 +1,1 @@
+Content of the page. All examples are into PageArea section
